requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-openai==0.2.0
pydantic==2.9.0
python-dotenv==1.0.1
